+++ src/protocol/gossip/impl/stream.cpp
@@ -79,7 +79,7 @@ namespace libp2p::protocol::gossip {
 
     read_buffer_->resize(msg_len);
 
-    stream_->read(std::span(read_buffer_->data(), ssize_t(msg_len)), msg_len,
+    stream_->read(std::span(read_buffer_->data(), msg_len), msg_len,
                   [self_wptr = weak_from_this(), this,
                    buffer = read_buffer_](auto &&res) {
                     if (self_wptr.expired()) {
